requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-crispy-forms==1.8.0
django-feedparser==0.2.1
django-import-export==1.2.0
django-jet==1.0.8
et-xmlfile==1.0.1
feedparser==5.2.1
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
urllib3==1.25.7
xlrd==1.2.0
xlwt==1.3.0
